modify serviceController valid port

+++ src/main/java/com/bonc/epm/paas/controller/ServiceController.java
@@ -4,6 +4,8 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -76,7 +78,7 @@ import com.github.dockerjava.api.model.ExposedPort;
 public class ServiceController {
 	private static final Logger log = LoggerFactory.getLogger(ServiceController.class);
 
-	Set<Integer> smalSet;
+	public static Set<Integer> smalSet = new HashSet<Integer>();
 	@Autowired
 	public ServiceDao serviceDao;
 	
@@ -441,7 +443,12 @@ public class ServiceController {
                 for (int i = 0;i<countOfExposedPort;i++) {
                     PortConfig portConfig = new PortConfig();
                     portConfig.setContainerPort(String.valueOf(exposedPorts[i].getPort()));
-                    portConfig.setMapPort(String.valueOf(vailPortSet()));
+                    int randomPort = vailPortSet();
+                    if (-1 != randomPort) {
+                        portConfig.setMapPort(String.valueOf(randomPort));                       
+                    } else {
+                        new RuntimeException("获取端口失败！");
+                    }
                     tmpPortConfigs.add(portConfig);
                 }
                 return tmpPortConfigs;
@@ -453,8 +460,7 @@ public class ServiceController {
     @RequestMapping(value = { "service/getPortConfig.do" }, method = RequestMethod.GET)
     @ResponseBody
     public String getPortConfig(String imgID){
-        List<PortConfig> list = new ArrayList();
-        list=getBaseImageExposedPorts(imgID);
+        List<PortConfig> list=getBaseImageExposedPorts(imgID);
         Map<String,Object> map = new HashMap<String,Object>();
         map.put("data",list);
         return JSON.toJSONString(map); 
@@ -503,11 +509,11 @@ public class ServiceController {
 		long userId = CurrentUserUtils.getInstance().getUser().getId();
 		Map<String, Object> map = new HashMap<String, Object>();
 		List<Image> images = imageDao.findAll(userId);
-		if (CollectionUtils.isNotEmpty(images)) {
+/*		if (CollectionUtils.isNotEmpty(images)) {
 		    for (Image image : images) {
 		        image.setPortConfigs(getBaseImageExposedPorts(String.valueOf(image.getId())));
 		    }
-		}
+		}*/
 		map.put("data", images);
 		System.out.println(JSON.toJSONString(map));
 		return JSON.toJSONString(map);
@@ -524,11 +530,11 @@ public class ServiceController {
 		long userId = CurrentUserUtils.getInstance().getUser().getId();
 		Map<String, Object> map = new HashMap<String, Object>();
 		List<Image> images = imageDao.findByNameOf(userId, "%" + imageName + "%");
-		if (CollectionUtils.isNotEmpty(images)) {
+/*		if (CollectionUtils.isNotEmpty(images)) {
 		    for (Image image : images) {
 		        image.setPortConfigs(getBaseImageExposedPorts(String.valueOf(image.getId())));
 		    }
-		}
+		}*/
 		map.put("data", images);
 		return JSON.toJSONString(map);
 	}
@@ -659,6 +665,9 @@ public class ServiceController {
 				portCon.setCreateDate(new Date());
 				portCon.setServiceId(service.getId());
 				portConfigDao.save(portCon);
+				
+				// 向map中添加生成的node端口
+				smalSet.add(Integer.valueOf(portCon.getContainerPort()));
 			}
 		}
 		//保存
@@ -756,42 +765,47 @@ public class ServiceController {
 		map.put("data", envVariables);
 		return JSON.toJSONString(map);
 	}
-	
 	/**
-	 * yuanpeng
-	 * 生成有效的PORTSET
-	 * @return int
-	 */
-	public int vailPortSet(){
-		int offset = kubernetesClientService.getK8sEndPort() - kubernetesClientService.getK8sStartPort();
-		Set<Integer> bigSet = Stream.iterate(kubernetesClientService.getK8sStartPort(), item -> item+1)
-									.limit(offset)
-									.collect(Collectors.toSet());
-		if(CollectionUtils.isEmpty(smalSet)){
-			smalSet= portConfigDao.findPortSets();
-			smalSet.remove(null);//
-		}	 else{
-			bigSet.removeAll(smalSet);
-		}
-		if(CollectionUtils.isEmpty(bigSet)){
-			return -1;
-		}else{
-	//	int portSet=Integer.valueOf(obj[(int)(Math.random()*obj.length)].toString());
-			Object[] obj =bigSet.toArray();
-		int portSet=Integer.valueOf(obj[0]
-							.toString());
-		smalSet.add(portSet);
-		log.info("大小："+smalSet.size());
-		log.info(smalSet.toString());
-		return portSet;
-		}
-	}
+    * 生成有效的PORTSET,回收端口
+    * @return int
+    */
+   public int vailPortSet(){
+       synchronized (this) {
+           int offset = kubernetesClientService.getK8sEndPort() - kubernetesClientService.getK8sStartPort();
+           Set<Integer> bigSet = Stream.iterate(kubernetesClientService.getK8sStartPort(), item -> item+1)
+                                       .limit(offset)
+                                       .collect(Collectors.toSet());
+/*           Set<Integer> tmpBigSet = bigSet;
+           tmpBigSet.removeAll(smalSet);
+           if (CollectionUtils.isEmpty(tmpBigSet)) {
+               smalSet.clear();
+               smalSet.addAll(portConfigDao.findPortSets());
+               smalSet.remove(null);
+               bigSet.removeAll(smalSet); 
+               return -1;
+           } else {
+               smalSet.addAll(portConfigDao.findPortSets());
+               smalSet.remove(null);
+               bigSet.removeAll(smalSet);
+           }*/
+           bigSet.removeAll(smalSet);
+           if (CollectionUtils.isEmpty(bigSet)) {
+               smalSet.addAll(portConfigDao.findPortSets());
+               smalSet.remove(null);
+               return -1;
+           }
+           Object[] obj =bigSet.toArray();
+           int portSet=Integer.valueOf(obj[(int)(Math.random()*obj.length)]
+                              .toString());
+           return portSet;
+       }
+   }
 
 	@RequestMapping(value = {"service/generatePortSet.do"} , method = RequestMethod.GET)
 	@ResponseBody
 	public String generatePortSet(){
 		Map<String, String> map = new HashMap<String, String>();
-		if(-1==vailPortSet()){
+		if(-1 == vailPortSet()){
 			map.put("ERROR","error");
 			return JSON.toJSONString(map);
 		}
@@ -1062,7 +1076,16 @@ public class ServiceController {
 			map.put("status", "200");
 			serviceDao.delete(id);
 			envVariableDao.deleteByServiceId(id);
-			portConfigDao.deleteByServiceId(id);
+			
+			// 删除服务 释放绑定的端口
+			List<PortConfig> bindPort = portConfigDao.findByServiceId(id);
+			if (CollectionUtils.isNotEmpty(bindPort)) {
+			    for (PortConfig oneRow : bindPort) {
+			        smalSet.remove(Integer.valueOf(oneRow.getContainerPort()));
+			    }
+			    portConfigDao.deleteByServiceId(id);
+			}
+			
 			// 更新挂载卷的使用状态
 			this.updateStorageType(service.getVolName(), service.getServiceName());
 		} catch (KubernetesClientException e) {
